backend/requirements.txt
asgiref==3.10.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.8
django-appconf==1.2.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
pillow==12.0.0
pycparser==2.23
razorpay==2.0.0
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
